update code use text to find element
updated

+++ src/test/java/Util/Configuration.java
@@ -1,18 +1,24 @@
 package Util;
 
+import com.codeborne.selenide.SelenideConfig;
+
+import java.sql.Time;
+
 public class Configuration {
     public static String WEB_BROWSER = "";
-    public static int TIME_OUT= 0;
+    public static long  TIME_OUT= 0;
     public static int PAGE_LOAD_TIME= 0;
     public static boolean DEFAULT_MAXIMUM =true;
+//    private static final SelenideConfig defaults = new SelenideConfig();
 
     public static void  ReadConfig(){
         if(CheckParameter("WEB_BROWSER")){
             Configuration.WEB_BROWSER = System.getenv("WEB_BROWSER");
         }
-        Configuration.TIME_OUT = Integer.parseInt(System.getProperty("timeout"));
+        Configuration.TIME_OUT = Long.parseLong(System.getProperty("timeout"));
         Configuration.PAGE_LOAD_TIME = Integer.parseInt(System.getProperty("pageLoadTimeout"));
         Configuration.DEFAULT_MAXIMUM = Boolean.parseBoolean(System.getProperty("startMaximized"));
+//        TIME_OUT = defaults.timeout();
     }
     public static boolean CheckParameter(String para){
         if(System.getenv(para)!=null){
@@ -20,4 +26,5 @@ public class Configuration {
         }
         return false;
     }
+
 }
